DOST: Distance-preserving Optimization for Spatial Transcriptomics

DOST is a distance-preserving optimization based method that clusters spatial transcriptomics (ST) spots into spatial domains. DOST embeds ST spots into a low-dimensional space using a Sammon mapping inspired loss function that jointly preserves gene expression similarity and spatial smoothness.

See also

Author

Maintainer: Mısra Taşçı mtasci18@ku.edu.tr

Authors:

Mehmet Gönen mehmetgonen@ku.edu.tr